major revision to installed scripts, using entry_points and apps.py instead of scripts in bin

+++ setup.py
@@ -12,11 +12,17 @@ import site
 import platform
 from glob import glob
 import shutil
+import subprocess
 
 DEBUG = False
 cmdline_args = sys.argv[1:]
 INSTALL =  len(cmdline_args)> 0 and (cmdline_args[0] == 'install')
 
+
+nbits = platform.architecture()[0].replace('bit', '')
+uname = sys.platform
+if os.name == 'nt':
+    uname = 'win'
 
 _version__ = None
 with open(os.path.join('lib', 'version.py'), 'r') as version_file:
@@ -30,7 +36,6 @@ with open(os.path.join('lib', 'version.py'), 'r') as version_file:
 
 ##
 ## Dependencies: required and recommended modules
-
 required_modules = {'numpy': 'numpy',
                     'scipy': 'scipy',
                     'matplotlib': 'matplotlib',
@@ -39,6 +44,7 @@ required_modules = {'numpy': 'numpy',
                     'requests': 'requests',
                     'six' : 'six',
                     'psutil': 'psutil',
+                    'pyshortcuts': 'pyshortcuts',
                     'peakutils': 'peakutils',
                     'PIL' : 'pillow',
                     'asteval': 'asteval',
@@ -48,15 +54,12 @@ required_modules = {'numpy': 'numpy',
                     'termcolor': 'termcolor'}
 
 
-graphics_modules = {'wx': 'wxPython',
-                    'wxmplot': 'wxmplot',
-                    'wxutils': 'wxutils'}
+graphics_modules = {'wx': 'wxPython', 'wxmplot': 'wxmplot', 'wxutils':'wxutils'}
 
-xrd_modules  = {'pyFAI': 'pyFAI', 'CifFile' : 'PyCifRW',
-                'fabio': 'fabio'}
+xrd_modules  = {'pyFAI': 'pyFAI', 'CifFile' : 'PyCifRW', 'fabio': 'fabio',
+                'dioptas': 'Dioptas'}
 
-tomo_modules = {'tomopy': 'tomopy',
-                'skimage': 'scikit-image'}
+tomo_modules = {'tomopy': 'tomopy', 'skimage': 'scikit-image'}
 
 epics_modules = {'epics': 'pyepics'}
 scan_modules = {'epicsscan': 'epicsscan', 'psycopg2': 'psycopg2'}
@@ -73,25 +76,22 @@ all_modules = (('basic analysis', required_modules),
                ('connecting to the EPICS control system', epics_modules),
                ('reading Spec files', spec_modules),
                ('PCA and machine learning', pca_modules),
-               # ('scanning with EpicsScan', scan_modules),
                ('testing tools',  testing_modules))
 
 
-# files that may be left from earlier install(s) and should be removed
-historical_cruft = ['plugins/xrd/xrd_hkl.py',
-                    'plugins/xrd/xrd_util.py',
-                    'plugins/xrd/xrd_xutil.py',
-                    'bin/larch_makeicons']
+compiled_exes = ('feff6l', 'feff8l_ff2x', 'feff8l_genfmt',
+                 'feff8l_pathfinder', 'feff8l_pot', 'feff8l_rdinp',
+                 'feff8l_xsph')
 
 modules_imported = {}
 missing = []
-
 
 try:
     import matplotlib
     matplotlib.use('WXAgg')
 except:
     pass
+
 print( 'Checking dependencies....')
 for desc, mods in all_modules:
     for impname, modname in mods.items():
@@ -115,30 +115,26 @@ pjoin = os.path.join
 psplit = os.path.split
 pexists = os.path.exists
 
-
 # system-wide larch directory
 larchdir = pjoin(sys.exec_prefix, 'share', 'larch')
 
-##
 ## determine this platform for dlls and exes
-
 nbits = platform.architecture()[0].replace('bit', '')
 
-uname  = 'linux'
 libfmt = 'lib%s.so'
+exefmt = "%s"
 bindir = 'bin'
 pyexe = pjoin(bindir, 'python')
 
-if os.name == 'nt':
-    uname  = 'win'
+if uname == 'darwin':
+    libfmt = 'lib%s.dylib'
+elif uname == 'win':
     libfmt = '%s.dll'
+    exefmt = "%s.exe"
     bindir = 'Scripts'
     pyexe = 'python.exe'
-elif sys.platform == 'darwin':
-    uname  = 'darwin'
-    libfmt = 'lib%s.dylib'
 
-uname = "%s%s" % (uname, nbits)
+sname = "%s%s" % (uname, nbits)
 
 if DEBUG:
     print("##  Settings  (Debug mode) ## ")
@@ -150,24 +146,22 @@ if DEBUG:
 
 
 # construct list of files to install besides the normal python modules
-# this includes the larch executable files, and all the larch modules
-# and plugins
+# this includes the larch executable files, and all the larch plugins
 
-scripts  =  glob('bin/*')
-if not uname.startswith('win'):
-    scripts = [s for s in scripts if not s.endswith('.bat')]
-
-scripts.extend(glob("%s/*" % pjoin('exes', uname)))
-
-larch_mods = glob('modules/*.lar') + glob('modules/*.py')
 larch_icos = glob('icons/*.ic*')
+larch_exes = [pjoin('exes', sname, exefmt % exe) for exe in compiled_exes]
+larch_dlls = glob("%s/*" % pjoin('dlls', sname))
 
-larch_dlls = glob("%s/*" % pjoin('dlls', uname))
-
-data_files = [(bindir,  scripts),
+data_files = [(bindir, larch_exes),
               (pjoin(larchdir, 'icons'),       larch_icos),
-              (pjoin(larchdir, 'modules'),     larch_mods),
-              (pjoin(larchdir, 'dlls', uname), larch_dlls)]
+              (pjoin(larchdir, 'dlls', sname), larch_dlls)]
+
+
+scripts = ['larch', 'larch_server', 'feff8l', 'xas_viewer',
+           'gse_mapviewer', 'gse_dtcorrect', 'diffit1d', 'diffit2d',
+           'xrfdisplay', 'xrfdisplay_epics']
+
+larch_apps = ['{0:s} = larch.apps:run_{0:s}'.format(n) for n in scripts]
 
 plugin_dir = pjoin(larchdir, 'plugins')
 pluginfiles = []
@@ -197,45 +191,62 @@ for pdir in pluginpaths:
 with open('requirements.txt', 'r') as f:
     requirements = f.readlines()
 
+
 # now we have all the data files, so we can run setup
-_setup = setup(
-    name = 'xraylarch',
-    version = __version__,
-    author = 'Matthew Newville and the X-rayLarch Development Team',
-    author_email = 'newville@cars.uchicago.edu',
-    url          = 'http://xraypy.github.io/xraylarch/',
-    download_url = 'http://xraypy.github.io/xraylarch/',
-    license = 'BSD',
-    description = 'Synchrotron X-ray data analysis in python',
-    package_dir = {'larch': 'lib'},
-    packages = ['larch', 'larch.utils', 'larch.wxlib', 'larch.fitting'],
-    install_requires=requirements,
-    data_files  = data_files,
-    platforms = ['Windows', 'Linux', 'Mac OS X'],
-    classifiers=['Intended Audience :: Science/Research',
-                 'Operating System :: OS Independent',
-                 'Programming Language :: Python',
-                 'Topic :: Scientific/Engineering'],
-)
+setup(name = 'xraylarch',
+      version = __version__,
+      author = 'Matthew Newville and the X-rayLarch Development Team',
+      author_email = 'newville@cars.uchicago.edu',
+      url          = 'http://xraypy.github.io/xraylarch/',
+      download_url = 'http://xraypy.github.io/xraylarch/',
+      license = 'BSD',
+      description = 'Synchrotron X-ray data analysis in python',
+      package_dir = {'larch': 'lib'},
+      packages = ['larch', 'larch.utils', 'larch.wxlib', 'larch.fitting'],
+      install_requires=requirements,
+      data_files  = data_files,
+      entry_points = {'gui_scripts' : larch_apps},
+      platforms = ['Windows', 'Linux', 'Mac OS X'],
+      classifiers=['Intended Audience :: Science/Research',
+                   'Operating System :: OS Independent',
+                   'Programming Language :: Python',
+                   'Topic :: Scientific/Engineering'],
+      )
 
+print("# setup done, finishing installation")
 
-def remove_cruft(basedir, filelist):
-    """remove files from base directory"""
+def remove_cruft():
+    """remove files that may be left from earlier installs"""
+    cruft = {'plugins': ['xrd/xrd_hkl.py',
+                         'xrd/xrd_util.py',
+                         'xrd/xrd_xutil.py'],
+             'bin': ['larch_makeicons', 'larch_gui', 'larch_client',
+                     'gse_scanviewer']}
+
     def remove_file(base, fname):
         fullname = pjoin(base, fname)
         if pexists(fullname):
-            print( "should remove %s " %  fullname)
             try:
-                print(os.stat(fullname))
-                # os.unlink(fullname)
+                os.unlink(fullname)
             except:
                 pass
 
-    for fname in filelist:
-        remove_file(basedir, fname)
-        if fname.endswith('.py'):
-            remove_file(basedir, fname+'c')
-            remove_file(basedir, fname+'o')
+    for category, flist in cruft.items():
+        if category == 'plugins':
+            basedir = pjoin(larchdir, 'plugins')
+            for fname in flist:
+                remove_file(basedir, fname)
+                if fname.endswith('.py'):
+                    remove_file(basedir, fname+'c')
+                    remove_file(basedir, fname+'o')
+
+        if category == 'bin':
+            basedir = pjoin(sys.exec_prefix, bindir)
+            for fname in flist:
+                remove_file(basedir, fname)
+                if fname.endswith('.py'):
+                    remove_file(basedir, fname+'c')
+                    remove_file(basedir, fname+'o')
 
 def remove_distutils_sitepackage():
     """rename site-packages/larch folder that may be
@@ -244,6 +255,8 @@ def remove_distutils_sitepackage():
         lpath = os.path.join(spath, 'larch')
         if os.path.exists(lpath) and os.path.isdir(lpath):
             dest = lpath + '_outdated'
+            if os.path.exists(dest):
+                shutil.rmtree(dest)
             try:
                 shutil.move(lpath, dest)
             except:
@@ -258,16 +271,14 @@ def fix_darwin_dylibs():
     larchdlls = 'share/larch/dlls/darwin64'
 
     dylibs = ('libgcc_s.1.dylib','libquadmath.0.dylib', 'libgfortran.3.dylib',
+
               'libfeff6.dylib', 'libcldata.dylib', 'libfeff8lpath.dylib',
               'libfeff8lpotph.dylib')
-
-    exes = ('feff6l', 'feff8l_ff2x', 'feff8l_genfmt', 'feff8l_pathfinder',
-            'feff8l_pot', 'feff8l_rdinp', 'feff8l_xsph')
 
     fixcmd = '/usr/bin/install_name_tool -change'
 
     cmds = []
-    for ename in exes:
+    for ename in compiled_exes:
         ename = pjoin(newdir, 'bin', ename)
         for dname in dylibs:
             old = pjoin(olddir, dname)
@@ -284,6 +295,27 @@ def fix_darwin_dylibs():
     for cmd in cmds:
         os.system(cmd)
 
+def fix_darwin_exes():
+    "fix anaconda python apps on MacOs to launch with pythonw"
+
+    if 'Anaconda' not in sys.version:
+        return
+    pyapp = pjoin(sys.prefix, 'python.app', 'Contents', 'MacOS', 'python')
+    for script in scripts:
+        appname = os.path.join(sys.exec_prefix, 'bin', script)
+        if os.path.exists(appname):
+            with open(appname, 'r') as fh:
+                try:
+                    lines = fh.readlines()
+                except IOError:
+                    lines = ['-']
+            time.sleep(.025)
+            if len(lines) > 1:
+                text = ["#!%s\n" % pyapp]
+                text.extend(lines[1:])
+                with open(appname, 'w') as fh:
+                    fh.write("".join(text))
+
 def fix_linux_dylibs():
     """
     fix dynamic libs on Linux with patchelf
@@ -297,46 +329,31 @@ def fix_linux_dylibs():
               'libfeff6.so', 'libcldata.so', 'libfeff8lpath.so',
               'libfeff8lpotph.so')
 
-    exes = ('feff6l', 'feff8l_ff2x', 'feff8l_genfmt', 'feff8l_pathfinder',
-            'feff8l_pot', 'feff8l_rdinp', 'feff8l_xsph')
 
     for lname in dylibs:
         os.system("%s '$ORIGIN' %s" % (fixcmd, os.path.join(larchdlls, lname)))
 
-    for ename in exes:
+    for ename in compiled_exes:
         os.system("%s %s %s/bin/%s" % (fixcmd, larchdlls, prefix, ename))
 
+# on install:
+#   remove historical cruft
+#   fix dynamic libraries
+#   fix MacOS + Anaconda python vs. pythonw
+#   create desktop icons
 if INSTALL:
-    remove_cruft(larchdir, historical_cruft)
+    remove_cruft()
     remove_distutils_sitepackage()
 
-    scriptdir = pjoin(sys.exec_prefix, bindir)
-    install_prefix = _setup.get_command_obj('install').root
-    if install_prefix is not None:
-        scriptdir = pjoin(install_prefix, scriptdir)
-
-    for src in scripts:
-        _, fname = psplit(src)
-        dest = pjoin(scriptdir, fname)
-        try:
-            shutil.copy(src, dest)
-            os.chmod(dest, 493) # mode=755
-        except:
-            pass
-
-# final install:
-#   create desktop icons
-#   fix dynamic libraries
-if INSTALL:
     if uname.startswith('darwin'):
         fix_darwin_dylibs()
+        if 'Anaconda' in sys.version:
+            fix_darwin_exes()
     elif uname.startswith('linux'):
         fix_linux_dylibs()
 
-    print( " could run larch makeicons")
-    # cmd ="%s %s" % (pjoin(sys.exec_prefix, pyexe),
-    #                 pjoin(sys.exec_prefix, bindir, 'larch_makeicons'))
-    # os.system(cmd)
+    subprocess.check_call((pjoin(sys.exec_prefix, pyexe),
+                           pjoin(sys.exec_prefix, bindir, 'larch'), '-m'))
